Turn prettynice/lib into a package at root of repo
It's a node package, even though there is some client-side code needed.
Can't make it a common package since there's a hard dependency on node.
So for now it's dropping the little bit of code needed on the frontend
into a .prettynice directory (where it's also putting the generated code
for client-side components).