Expose low-level functions to for program records
Went back and forth on whether to call define/start program for you or
give you an API to do it yourself (explicitness, safety, and guarantees
preference for the latter, but huge usability preference, especially for
beginners, for the first). So punting on this decision and exposing both
for now.